Update and rename BubbleSort.cpp to SelectionSort.cpp
